wpt_babel/content-security-policy/securitypolicyviolation/img-src-redirect.sub.html

<!DOCTYPE html>
<meta content="img-src 'self'" http-equiv="content-security-policy"/>
<script src="/resources/testharness.js">
</script>
<script src="/resources/testharnessreport.js">
</script>
<body>
</body>
<script>
 "use strict";

async_test(function (t) {
  var i = document.createElement("img");
  var target = "http://{{hosts[alt][]}}:{{ports[http][0]}}/content-security-policy/support/fail.png";
  var url = window.origin + "/common/redirect.py?location=" + encodeURIComponent(target);
  window.addEventListener('securitypolicyviolation', t.step_func_done(function (e) {
    assert_equals(e.blockedURI, url);
  }));
  i.onload = t.step_func(function () {
    assert_unreached("Img should be blocked.");
  });
  i.src = url;
  document.body.appendChild(i);
}, "The blocked URI in the security policy violation event should be the original URI before redirects.");
</script>
